requirements.txt
matplotlib
scikit-learn
Pillow==9.4.0
scikit-plot==0.3.7
seaborn
numpy
pandas
xgboost
lightgbm
pyradiomics
SimpleITK